requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
drf-spectacular>=0.26.5
django-filter>=23.5
python-json-logger>=2.0.7

# Database
dj-database-url>=2.1.0

# Authentication (opcional)
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Linting & Formatting
ruff>=0.1.6
black>=23.11.0

# Pre-commit
pre-commit>=3.5.0

# Observability
prometheus-client>=0.19.0

# Optional: Sentry
# sentry-sdk>=1.38.0

